Try to improve the reliability of the ANSICON installation.
NUSPEC

Update the version to 1.66. There is no patch number and the minor number is
definitely two digits. The release page, archive, and help display the same

  > ansicon /?
  ANSICON by Jason Hood <jadoxa@yahoo.com.au>.
  Version 1.66 (21 September, 2013).  Freeware.

BUGS

Older versions of this package had a few inconsistencies that may have caused
issues when installing or upgrading over a "proper" install (`ansicon -i`).
This version should more accurately update proper installs, older versions of
this package, or hand-edited commands.

  * ANSICON uses a single `&` between its command and any existing AutoRun
    commands. This command separator ignores exit codes. The double `&&` does
    logical-OR shortcutting.

  * ANSICON does not include the file extension (.exe) in the command.

SCRIPT

Download and upack the archive in the Chocolatey package directory, not in the
%TEMP% dir. Use the built-in Chocolatey command to get the processor bitness.
This greatly simplifies that part of the script.

Use a regex to pull the exact ANSICON command from the AutoRun property. This
implementation is only 2 levels deep, the previous was 5 levels deep in two
separate paths). This also preserves any user customizations.